requirements.txt
aniso8601==8.0.0
appdirs==1.4.4
asgiref==3.3.4
certifi==2020.12.5
click==7.1.2
coverage==5.5
distlib==0.3.1
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
drf-nested-routers==0.93.3
filelock==3.0.12
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
greenlet==1.0.0
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
Markdown==3.3.4
MarkupSafe==1.1.1
Pillow==8.2.0
pipenv==2020.11.15
psycopg2==2.8.6
PyJWT==2.0.1
python-dotenv==0.17.0
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
sqlparse==0.4.1
virtualenv==20.4.2
virtualenv-clone==0.5.4
Werkzeug==1.0.1
whitenoise==5.2.0
